further decoupling: - introduced backend classes - implement basic matplotlib - altair is very unfinished
