+++ mesa/visualization/backends/__init__.py
@@ -1,0 +1,7 @@
+from .altair_backend import AltairBackend
+from .matplotlib_backend import MatplotlibBackend
+
+__all__ = [
+    "AltairBackend",
+    "MatplotlibBackend",
+]
