+++ mesa/visualization/space_renderer.py
@@ -28,21 +28,18 @@ Classes:
     SpaceRenderer: Main rendering class for Mesa spaces
 """
 
-import os
 import warnings
-from collections.abc import Callable
-from dataclasses import fields
 
-import altair as alt
 import numpy as np
-import pandas as pd
-from matplotlib import pyplot as plt
 from matplotlib.axes import Axes
-from matplotlib.cm import ScalarMappable
-from matplotlib.collections import PolyCollection
-from matplotlib.colors import LinearSegmentedColormap, Normalize, to_rgb, to_rgba
-from matplotlib.offsetbox import AnnotationBbox, OffsetImage
-from PIL import Image
+from solara import Literal
+from space_drawers import (
+    ContinuousSpaceDrawer,
+    HexSpaceDrawer,
+    NetworkSpaceDrawer,
+    OrthogonalSpaceDrawer,
+    VoronoiSpaceDrawer,
+)
 
 import mesa
 from mesa.discrete_space import (
@@ -58,36 +55,25 @@ from mesa.space import (
     NetworkGrid,
     SingleGrid,
 )
-from mesa.visualization.space_drawers import (
-    ContinuousSpaceDrawer,
-    HexSpaceDrawer,
-    NetworkSpaceDrawer,
-    OrthogonalSpaceDrawer,
-    VoronoiSpaceDrawer,
-)
+from mesa.visualization.backends import AltairBackend, MatplotlibBackend
 
-CORRECTION_FACTOR_MARKER_ZOOM = 0.01
-
+# Define type hints for space types
 OrthogonalGrid = SingleGrid | MultiGrid | OrthogonalMooreGrid | OrthogonalVonNeumannGrid
 HexGrid = HexSingleGrid | HexMultiGrid | mesa.discrete_space.HexGrid
 Network = NetworkGrid | mesa.discrete_space.Network
 
 
 class SpaceRenderer:
-    """Renders Mesa spaces with agents and property layers using different backends."""
+    """Renders Mesa spaces using different backends."""
 
-    def __init__(self, model, backend: str = "matplotlib", ax: Axes | None = None):
-        """Initialize the space renderer.
+    def __init__(
+        self,
+        model,
+        backend: Literal["matplotlib", "altair"] | None = "matplotlib",
+        **kwargs,
+    ):
+        self.space = getattr(model, "grid", getattr(model, "space", None))
 
-        Args:
-            model: Mesa model containing the space to render
-            backend: Rendering backend ('matplotlib' or 'altair')
-            ax: Matplotlib axes (ignored for altair backend)
-        """
-        self.space = getattr(model, "grid", None)
-        if self.space is None:
-            self.space = getattr(model, "space", None)
-        self.backend = backend
         self.space_drawer = self._get_space_drawer()
 
         # Initialize mesh storage
@@ -95,21 +81,23 @@ class SpaceRenderer:
         self.agent_mesh = None
         self.propertylayer_mesh = None
 
-        # Handle axis management
-        if self.backend == "matplotlib":
-            if ax is None:
-                # constrained_layout=True is must because it adjusts the layout automatically
-                # to prevent shrinking of the figure due to non-existent colorbar.
-                fig, ax = plt.subplots(constrained_layout=True)
-                self.fig = fig
-            self.ax = ax
-        else:
-            self.ax = None
-            warnings.warn(
-                "Altair backend does not support direct axis management. ",
-                UserWarning,
-                stacklevel=2,
+        self.backend = backend
+
+        if backend == "matplotlib":
+            self.backend_renderer = MatplotlibBackend(
+                self.space_drawer,
+                [self.space_mesh, self.agent_mesh, self.propertylayer_mesh],
             )
+            self.canvas = self.backend_renderer.canvas
+        elif backend == "altair":
+            self.backend_renderer = AltairBackend(
+                self.space_drawer,
+                [self.space_mesh, self.agent_mesh, self.propertylayer_mesh],
+            )
+        else:
+            raise ValueError(f"Unsupported backend: {backend}")
+
+        self.backend_renderer.initialize_canvas(**kwargs)
 
     def _get_space_drawer(self):
         """Get appropriate space drawer based on space type."""
@@ -164,834 +152,31 @@ class SpaceRenderer:
 
         return mapped_arguments
 
-    # Structure drawing methods
     def draw_structure(self, ax: Axes | None = None, **kwargs):
-        """Draw the space structure (grid lines, etc.)."""
-        self.space_kwargs = kwargs.copy()
-        if self.backend == "matplotlib":
-            ax = ax if ax is not None else self.ax
-            self.space_mesh = self.space_drawer.draw_matplotlib(ax, **self.space_kwargs)
-            return self.space_mesh
-        else:
-            self.space_mesh = self.space_drawer.draw_altair(**self.space_kwargs)
-            return self.space_mesh
+        self.space_kwargs = kwargs
 
-    # Agent data collection methods
-    def _get_agent_pos(self, agent, space):
-        """Get agent position based on space type."""
-        if isinstance(space, NetworkGrid):
-            return agent.pos, agent.pos
-        elif isinstance(space, Network):
-            return agent.cell.coordinate, agent.cell.coordinate
-        else:
-            x = agent.pos[0] if agent.pos is not None else agent.cell.coordinate[0]
-            y = agent.pos[1] if agent.pos is not None else agent.cell.coordinate[1]
-            return x, y
-
-    def _mpl_collect_agent_data(
-        self,
-        space: OrthogonalGrid | HexGrid | Network | ContinuousSpace | VoronoiGrid,
-        agent_portrayal: Callable,
-        default_size: float | None = None,
-    ) -> dict:
-        """Collect plotting data for all agents in the space.
-
-        Args:
-            space: The space containing the agents
-            agent_portrayal: Callable that returns AgentPortrayalStyle for each agent
-            default_size: Default marker size if not specified
-
-        Returns:
-            Dictionary containing agent plotting data arrays
-        """
-        # Initialize data collection arrays
-        arguments = {
-            "loc": [],
-            "s": [],
-            "c": [],
-            "marker": [],
-            "zorder": [],
-            "alpha": [],
-            "edgecolors": [],
-            "linewidths": [],
-        }
-
-        # Import here to prevent circular imports
-        from mesa.visualization.components import AgentPortrayalStyle
-
-        # Get default values from AgentPortrayalStyle
-        style_fields = {f.name: f.default for f in fields(AgentPortrayalStyle)}
-        class_default_size = style_fields.get("size")
-
-        for agent in space.agents:
-            portray_input = agent_portrayal(agent)
-
-            if isinstance(portray_input, dict):
-                warnings.warn(
-                    "Returning a dict from agent_portrayal is deprecated. "
-                    "Please return an AgentPortrayalStyle instance instead.",
-                    PendingDeprecationWarning,
-                    stacklevel=2,
-                )
-                # Handle legacy dict input
-                dict_data = portray_input.copy()
-                agent_x, agent_y = self._get_agent_pos(agent, space)
-
-                # Extract values with defaults
-                aps = AgentPortrayalStyle(
-                    x=agent_x,
-                    y=agent_y,
-                    size=dict_data.pop("size", style_fields.get("size")),
-                    color=dict_data.pop("color", style_fields.get("color")),
-                    marker=dict_data.pop("marker", style_fields.get("marker")),
-                    zorder=dict_data.pop("zorder", style_fields.get("zorder")),
-                    alpha=dict_data.pop("alpha", style_fields.get("alpha")),
-                    edgecolors=dict_data.pop("edgecolors", None),
-                    linewidths=dict_data.pop(
-                        "linewidths", style_fields.get("linewidths")
-                    ),
-                )
-
-                # Warn about unused keys
-                if dict_data:
-                    ignored_keys = list(dict_data.keys())
-                    warnings.warn(
-                        f"The following keys were ignored: {', '.join(ignored_keys)}",
-                        UserWarning,
-                        stacklevel=2,
-                    )
-            else:
-                aps = portray_input
-                # Set defaults if not provided
-                if aps.edgecolors is None:
-                    aps.edgecolors = aps.color
-                if aps.x is None and aps.y is None:
-                    aps.x, aps.y = self._get_agent_pos(agent, space)
-
-            # Collect agent data
-            arguments["loc"].append((aps.x, aps.y))
-
-            # Determine final size
-            size_to_collect = aps.size or default_size or class_default_size
-            arguments["s"].append(size_to_collect)
-            arguments["c"].append(aps.color)
-            arguments["marker"].append(aps.marker)
-            arguments["zorder"].append(aps.zorder)
-            arguments["alpha"].append(aps.alpha)
-            if aps.edgecolors is not None:
-                arguments["edgecolors"].append(aps.edgecolors)
-            arguments["linewidths"].append(aps.linewidths)
-
-        # Convert to numpy arrays
-        data = {
-            k: (np.asarray(v, dtype=object) if k == "marker" else np.asarray(v))
-            for k, v in arguments.items()
-        }
-
-        # Handle marker array specially to preserve tuples
-        arr = np.empty(len(arguments["marker"]), dtype=object)
-        arr[:] = arguments["marker"]
-        data["marker"] = arr
-
-        return data
-
-    def _altair_collect_agent_data(
-        self,
-        space: OrthogonalGrid | HexGrid | Network | ContinuousSpace | VoronoiGrid,
-        agent_portrayal: Callable,
-        default_size: float | None = None,
-    ) -> dict:
-        """Collect plotting data for all agents in the space for Altair.
-
-        Args:
-            space: The space containing the agents
-            agent_portrayal: Callable that returns AgentPortrayalStyle for each agent
-            default_size: Default marker size if not specified
-
-        Returns:
-            Dictionary containing agent plotting data arrays
-        """
-        # Initialize data collection arrays
-        arguments = {
-            "loc": [],
-            "size": [],
-            "color": [],
-            "shape": [],
-            "order": [],  # z-order
-            "opacity": [],
-            "stroke": [],  # Stroke color
-            "strokeWidth": [],
-            "filled": [],
-        }
-
-        from mesa.visualization.components import AgentPortrayalStyle
-
-        style_fields = {f.name: f.default for f in fields(AgentPortrayalStyle)}
-        class_default_size = style_fields.get("size")
-
-        # Marker mapping from Matplotlib to Altair
-        marker_to_shape_map = {
-            "o": "circle",
-            "s": "square",
-            "D": "diamond",
-            "^": "triangle-up",
-            "v": "triangle-down",
-            "<": "triangle-left",
-            ">": "triangle-right",
-            "+": "cross",
-            "x": "cross",  # Both '+' and 'x' map to cross in Altair
-            ".": "circle",  # Small point becomes circle
-            "1": "triangle-down",
-            "2": "triangle-up",
-            "3": "triangle-left",
-            "4": "triangle-right",
-        }
-
-        for agent in space.agents:
-            portray_input = agent_portrayal(agent)
-            aps: AgentPortrayalStyle
-
-            if isinstance(portray_input, dict):
-                warnings.warn(
-                    "Returning a dict from agent_portrayal is deprecated. "
-                    "Please return an AgentPortrayalStyle instance instead.",
-                    PendingDeprecationWarning,
-                    stacklevel=2,
-                )
-                dict_data = portray_input.copy()
-                agent_x, agent_y = self._get_agent_pos(agent, space)
-
-                aps = AgentPortrayalStyle(
-                    x=agent_x,
-                    y=agent_y,
-                    size=dict_data.pop("size", style_fields.get("size")),
-                    color=dict_data.pop("color", style_fields.get("color")),
-                    marker=dict_data.pop("marker", style_fields.get("marker")),
-                    zorder=dict_data.pop("zorder", style_fields.get("zorder")),
-                    alpha=dict_data.pop("alpha", style_fields.get("alpha")),
-                    edgecolors=dict_data.pop(
-                        "edgecolors", style_fields.get("edgecolors")
-                    ),
-                    linewidths=dict_data.pop(
-                        "linewidths", style_fields.get("linewidths")
-                    ),
-                )
-                if dict_data:
-                    ignored_keys = list(dict_data.keys())
-                    warnings.warn(
-                        f"The following keys were ignored from dict portrayal: {', '.join(ignored_keys)}",
-                        UserWarning,
-                        stacklevel=2,
-                    )
-            else:
-                aps = portray_input
-                if aps.x is None and aps.y is None:
-                    aps.x, aps.y = self._get_agent_pos(agent, space)
-
-            arguments["loc"].append((aps.x, aps.y))
-
-            size_to_collect = aps.size if aps.size is not None else default_size
-            if size_to_collect is None:
-                size_to_collect = class_default_size
-            arguments["size"].append(size_to_collect)
-
-            arguments["color"].append(
-                aps.color if aps.color is not None else style_fields.get("color")
-            )
-
-            # Map marker to Altair shape if defined, else use raw marker
-            raw_marker = (
-                aps.marker if aps.marker is not None else style_fields.get("marker")
-            )
-            shape_value = marker_to_shape_map.get(raw_marker, raw_marker)
-            if shape_value is None:
-                warnings.warn(
-                    f"Marker '{raw_marker}' is not supported in Altair. "
-                    "Using 'circle' as default.",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                shape_value = "circle"
-            arguments["shape"].append(shape_value)
-
-            arguments["order"].append(
-                aps.zorder if aps.zorder is not None else style_fields.get("zorder")
-            )
-            arguments["opacity"].append(
-                aps.alpha if aps.alpha is not None else style_fields.get("alpha")
-            )
-            arguments["stroke"].append(aps.edgecolors)
-            arguments["strokeWidth"].append(
-                aps.linewidths
-                if aps.linewidths is not None
-                else style_fields.get("linewidths")
-            )
-
-            # FIXME: Kind of stupid logic.
-            # filled: True if edgecolors are defined, False otherwise.
-            filled_value = aps.edgecolors is not None
-            arguments["filled"].append(filled_value)
-
-        final_data = {}
-        for k, v in arguments.items():
-            if k == "shape":
-                # Ensure shape is an object array
-                arr = np.empty(len(v), dtype=object)
-                arr[:] = v
-                final_data[k] = arr
-            elif k in ["x", "y", "size", "order", "opacity", "strokeWidth"]:
-                final_data[k] = np.asarray(v, dtype=float)
-            else:
-                final_data[k] = np.asarray(v)
-
-        return final_data
-
-    # Agent drawing methods
-    def _mpl_draw_agents(self, ax: Axes | None, arguments, **kwargs):
-        """Draw agents using matplotlib backend."""
-
-        def _get_zoom_factor(ax, img):
-            """Calculate zoom factor for image markers based on axis limits."""
-            ax.get_figure().canvas.draw()
-            bbox = ax.get_window_extent().transformed(
-                ax.get_figure().dpi_scale_trans.inverted()
-            )  # in inches
-            width, height = (
-                bbox.width * ax.get_figure().dpi,
-                bbox.height * ax.get_figure().dpi,
-            )  # in pixel
-
-            xr = ax.get_xlim()
-            yr = ax.get_ylim()
-
-            x_pixel_per_data = width / (xr[1] - xr[0])
-            y_pixel_per_data = height / (yr[1] - yr[0])
-
-            zoom_x = (x_pixel_per_data / img.width) * CORRECTION_FACTOR_MARKER_ZOOM
-            zoom_y = (y_pixel_per_data / img.height) * CORRECTION_FACTOR_MARKER_ZOOM
-
-            return min(zoom_x, zoom_y)
-
-        loc = arguments.pop("loc")
-
-        loc_x = loc[:, 0]
-        loc_y = loc[:, 1]
-        marker = arguments.pop("marker")
-        zorder = arguments.pop("zorder")
-        malpha = arguments.pop("alpha")
-        msize = arguments.pop("s")
-
-        # we check if edgecolor, linewidth, and alpha are specified
-        # at the agent level, if not, we remove them from the arguments dict
-        # and fallback to the default value in ax.scatter / use what is passed via **kwargs
-        for entry in ["edgecolors", "linewidths"]:
-            if len(arguments[entry]) == 0:
-                arguments.pop(entry)
-            else:
-                if entry in kwargs:
-                    raise ValueError(
-                        f"{entry} is specified in agent portrayal and via plotting kwargs, you can only use one or the other"
-                    )
-
-        ax.get_figure().canvas.draw()
-        for mark in set(marker):
-            if isinstance(mark, (str | os.PathLike)) and os.path.isfile(mark):
-                # images
-                for m_size in np.unique(msize):
-                    image = Image.open(mark)
-                    im = OffsetImage(
-                        image,
-                        zoom=_get_zoom_factor(ax, image) * m_size,
-                    )
-                    im.image.axes = ax
-
-                    mask_marker = [m == mark for m in list(marker)] & (m_size == msize)
-                    for z_order in np.unique(zorder[mask_marker]):
-                        for m_alpha in np.unique(malpha[mask_marker]):
-                            mask = (
-                                (z_order == zorder) & (m_alpha == malpha) & mask_marker
-                            )
-                            for x, y in zip(loc_x[mask], loc_y[mask]):
-                                ab = AnnotationBbox(
-                                    im,
-                                    (x, y),
-                                    frameon=False,
-                                    pad=0.0,
-                                    zorder=z_order,
-                                    **kwargs,
-                                )
-                                ax.add_artist(ab)
-
-            else:
-                # ordinary markers
-                mask_marker = [m == mark for m in list(marker)]
-                for z_order in np.unique(zorder[mask_marker]):
-                    zorder_mask = z_order == zorder & mask_marker
-                    ax.scatter(
-                        loc_x[zorder_mask],
-                        loc_y[zorder_mask],
-                        marker=mark,
-                        zorder=z_order,
-                        **{k: v[zorder_mask] for k, v in arguments.items()},
-                        **kwargs,
-                    )
-
-    def _draw_agents_altair(
-        self, arguments, chart_width=450, chart_height=350, **kwargs
-    ):
-        """Draw agents using altair backend with customizable kwargs."""
-        if arguments["loc"].size == 0:
-            return None
-
-        # Agent data preparation
-        df_data = {
-            "x": arguments["loc"][:, 0],
-            "y": arguments["loc"][:, 1],
-            "size": arguments["size"],
-            "shape": arguments["shape"],
-            "opacity": arguments["opacity"],
-            "strokeWidth": arguments["strokeWidth"],
-            "original_color": arguments["color"],
-            "is_filled": arguments["filled"],
-            "original_stroke": arguments["stroke"],
-        }
-        df = pd.DataFrame(df_data)
-
-        # To ensure distinct shapes according to agnet portrayal
-        unique_shape_names_in_data = df["shape"].unique().tolist()
-
-        fill_colors = []
-        stroke_colors = []
-        for i in range(len(df)):
-            filled = df["is_filled"][i]
-            main_color = df["original_color"][i]
-            stroke_spec = (
-                df["original_stroke"][i]
-                if isinstance(df["original_stroke"][i], str)
-                else None
-            )
-            if filled:
-                fill_colors.append(main_color)
-                stroke_colors.append(stroke_spec)
-            else:
-                fill_colors.append(None)
-                stroke_colors.append(main_color)
-        df["viz_fill_color"] = fill_colors
-        df["viz_stroke_color"] = stroke_colors
-
-        # Extract additional parameters from kwargs
-        # FIXME: Add more parameters to kwargs
-        title = kwargs.pop("title", "")
-        xlabel = kwargs.pop("xlabel", "")
-        ylabel = kwargs.pop("ylabel", "")
-
-        # Tooltip list for interactivity
-        # FIXME: Add more fields to tooltip (preferably from agent_portrayal)
-        tooltip_list = ["x", "y"]
-
-        # Determine space dimensions
-        xmin, xmax, ymin, ymax = self.space_drawer.get_viz_limits()
-
-        chart = (
-            alt.Chart(df)
-            .mark_point()
-            .encode(
-                x=alt.X(
-                    "x:Q",
-                    title=xlabel,
-                    scale=alt.Scale(type="linear", domain=[xmin, xmax]),
-                ),
-                y=alt.Y(
-                    "y:Q",
-                    title=ylabel,
-                    scale=alt.Scale(type="linear", domain=[ymin, ymax]),
-                ),
-                size=alt.Size("size:Q", legend=None),
-                shape=alt.Shape(
-                    "shape:N",
-                    scale=alt.Scale(
-                        domain=unique_shape_names_in_data,
-                        range=unique_shape_names_in_data,
-                    ),
-                    title="Shape",
-                ),
-                opacity=alt.Opacity(
-                    "opacity:Q",
-                    title="Opacity",
-                    scale=alt.Scale(domain=[0, 1], range=[0, 1]),
-                ),
-                fill=alt.Fill("viz_fill_color:N", scale=None, title="Fill Color"),
-                stroke=alt.Stroke(
-                    "viz_stroke_color:N", scale=None, title="Stroke Color"
-                ),
-                strokeWidth=alt.StrokeWidth("strokeWidth:Q", title="Stroke Width"),
-                tooltip=tooltip_list,
-            )
-            .properties(title=title, width=chart_width, height=chart_height)
-        )
-
-        return chart
+        ax = ax or self.canvas
+        self.space_mesh = self.backend_renderer.draw_structure(ax, **self.space_kwargs)
+        return self.space_mesh
 
     def draw_agents(self, agent_portrayal, ax: Axes | None = None, **kwargs):
-        """Draw agents on the space."""
-        self.agent_kwargs = kwargs.copy()
-
         self.agent_portrayal = agent_portrayal
+        ax = ax or self.canvas
+        self.agent_kwargs = kwargs
+        arguments = self.backend_renderer._collect_agent_data(
+            self.space, agent_portrayal
+        )
+        arguments = self._map_coordinates(arguments)
+        self.agent_mesh = self.backend_renderer.draw_agents(
+            ax=ax, arguments=arguments, **self.agent_kwargs
+        )
+        return self.agent_mesh
 
-        if self.backend == "matplotlib":
-            ax = ax if ax is not None else self.ax
-
-            arguments = self._mpl_collect_agent_data(
-                self.space, agent_portrayal, default_size=self.space_drawer.s_default
-            )
-            arguments = self._map_coordinates(arguments)
-            self.agent_mesh = self._mpl_draw_agents(ax, arguments, **self.agent_kwargs)
-
-            return self.agent_mesh
-
-        else:
-            arguments = self._altair_collect_agent_data(
-                self.space, agent_portrayal, default_size=self.space_drawer.s_default
-            )
-            arguments = self._map_coordinates(arguments)
-            self.agent_mesh = self._draw_agents_altair(arguments, **self.agent_kwargs)
-            return self.agent_mesh
-
-    # Property layer drawing methods
-    def _draw_propertylayer_matplotlib(
-        self, space, property_layers, propertylayer_portrayal, ax
-    ):
-        """Draw property layers using matplotlib backend."""
-        # Draw each layer
-        for layer_name in property_layers:
-            if layer_name == "empty":
-                continue
-
-            layer = property_layers.get(layer_name)
-            portrayal = propertylayer_portrayal(layer)
-
-            if portrayal is None:
-                continue
-
-            data = layer.data.astype(float) if layer.data.dtype == bool else layer.data
-
-            # Check dimensions
-            if (space.width, space.height) != data.shape:
-                warnings.warn(
-                    f"Layer {layer_name} dimensions ({data.shape}) "
-                    f"don't match space dimensions ({space.width}, {space.height})",
-                    UserWarning,
-                    stacklevel=2,
-                )
-
-            # Get portrayal parameters
-            color = portrayal.color
-            colormap = portrayal.colormap
-            alpha = portrayal.alpha
-            vmin = portrayal.vmin if portrayal.vmin is not None else np.min(data)
-            vmax = portrayal.vmax if portrayal.vmax is not None else np.max(data)
-
-            # Set up colormap
-            if color:
-                rgba_color = to_rgba(color)
-                cmap = LinearSegmentedColormap.from_list(
-                    layer_name, [(0, 0, 0, 0), (*rgba_color[:3], alpha)]
-                )
-            elif colormap:
-                cmap = colormap
-                if isinstance(cmap, list):
-                    cmap = LinearSegmentedColormap.from_list(layer_name, cmap)
-                elif isinstance(cmap, str):
-                    cmap = plt.get_cmap(cmap)
-            else:
-                raise ValueError(
-                    f"PropertyLayer {layer_name} must include 'color' or 'colormap'"
-                )
-
-            # Draw based on space type
-            if isinstance(space, OrthogonalGrid):
-                if color:
-                    data = data.T
-                    normalized_data = (data - vmin) / (vmax - vmin)
-                    rgba_data = np.full((*data.shape, 4), rgba_color)
-                    rgba_data[..., 3] *= normalized_data * alpha
-                    rgba_data = np.clip(rgba_data, 0, 1)
-                    ax.imshow(rgba_data, origin="lower")
-                else:
-                    ax.imshow(
-                        data.T,
-                        cmap=cmap,
-                        alpha=alpha,
-                        vmin=vmin,
-                        vmax=vmax,
-                        origin="lower",
-                    )
-            elif isinstance(space, HexGrid):
-                hexagons = self.space_drawer.hexagons
-                norm = Normalize(vmin=vmin, vmax=vmax)
-                colors = data.ravel()
-
-                if color:
-                    normalized_colors = np.clip(norm(colors), 0, 1)
-                    rgba_colors = np.full((len(colors), 4), rgba_color)
-                    rgba_colors[:, 3] = normalized_colors * alpha
-                else:
-                    rgba_colors = cmap(norm(colors))
-                    rgba_colors[..., 3] *= alpha
-
-                collection = PolyCollection(hexagons, facecolors=rgba_colors, zorder=-1)
-                ax.add_collection(collection)
-            else:
-                raise NotImplementedError(
-                    f"PropertyLayer visualization not implemented for {type(space)}"
-                )
-
-            # Add colorbar if requested
-            cbar = None
-            if portrayal.colorbar:
-                norm = Normalize(vmin=vmin, vmax=vmax)
-                sm = ScalarMappable(norm=norm, cmap=cmap)
-                sm.set_array([])
-                cbar = plt.colorbar(sm, ax=ax, label=layer_name)
-
-        return ax, cbar
-
-    def _draw_propertylayer_altair(
-        self,
-        space,
-        property_layers,
-        propertylayer_portrayal,
-        chart_width=450,
-        chart_height=350,
-    ):
-        """Draw property layers using altair backend."""
-        base = None
-        bar_chart_viz = None
-
-        for layer_name in property_layers:
-            if layer_name == "empty":
-                continue
-
-            layer = property_layers.get(layer_name)
-            portrayal = propertylayer_portrayal(layer)
-
-            if portrayal is None:
-                continue
-
-            data = layer.data.astype(float) if layer.data.dtype == bool else layer.data
-
-            # Check dimensions
-            if (space.width, space.height) != data.shape:
-                warnings.warn(
-                    f"Layer {layer_name} dimensions ({data.shape}) do not match space dimensions ({space.width}, {space.height}).",
-                    UserWarning,
-                    stacklevel=2,
-                )
-                continue
-
-            # Get portrayal parameters
-            color = portrayal.color
-            colormap = portrayal.colormap
-            alpha = portrayal.alpha
-            vmin = portrayal.vmin if portrayal.vmin is not None else np.min(data)
-            vmax = portrayal.vmax if portrayal.vmax is not None else np.max(data)
-
-            # Prepare data for Altair
-            df = pd.DataFrame(
-                {
-                    "x": np.repeat(np.arange(data.shape[0]), data.shape[1]),
-                    "y": np.tile(np.arange(data.shape[1]), data.shape[0]),
-                    "value": data.flatten(),
-                }
-            )
-
-            current_chart = None
-            if color:
-                # Create a function to map values to RGBA colors with proper opacity scaling
-                def apply_rgba(
-                    val, v_min=vmin, v_max=vmax, a=alpha, p_color=portrayal.color
-                ):
-                    # Normalize value to range [0,1] and clamp
-                    normalized = max(
-                        0,
-                        min(
-                            ((val - v_min) / (v_max - v_min))
-                            if (v_max - v_min) != 0
-                            else 0.5,
-                            1,
-                        ),
-                    )
-
-                    # Scale opacity by alpha parameter
-                    opacity = normalized * a
-
-                    # Convert color to RGB components
-                    rgb_color_val = to_rgb(p_color)
-                    r = int(rgb_color_val[0] * 255)
-                    g = int(rgb_color_val[1] * 255)
-                    b = int(rgb_color_val[2] * 255)
-                    return f"rgba({r}, {g}, {b}, {opacity:.2f})"
-
-                # Apply color mapping to each value in the dataset
-                df["color_str"] = df["value"].apply(apply_rgba)
-
-                # Create chart for the property layer
-                current_chart = (
-                    alt.Chart(df)
-                    .mark_rect()
-                    .encode(
-                        x=alt.X("x:O", axis=None),
-                        y=alt.Y("y:O", axis=None),
-                        fill=alt.Fill("color_str:N", scale=None),
-                    )
-                    .properties(
-                        width=chart_width, height=chart_height, title=layer_name
-                    )
-                )
-                base = (
-                    alt.layer(current_chart, base)
-                    if base is not None
-                    else current_chart
-                )
-
-                # Add colorbar if specified in portrayal
-                if portrayal.colorbar:
-                    # Extract RGB components from base color
-                    rgb_color_val = to_rgb(portrayal.color)
-                    r_int = int(rgb_color_val[0] * 255)
-                    g_int = int(rgb_color_val[1] * 255)
-                    b_int = int(rgb_color_val[2] * 255)
-
-                    # Define gradient endpoints
-                    min_color_str = f"rgba({r_int},{g_int},{b_int},0)"
-                    max_color_str = f"rgba({r_int},{g_int},{b_int},{alpha:.2f})"
-
-                    # Define colorbar dimensions
-                    colorbar_height = 20
-                    colorbar_width = chart_width
-
-                    # Create dataframe for gradient visualization
-                    df_gradient = pd.DataFrame({"x_grad": [0, 1], "y_grad": [0, 1]})
-
-                    # Create evenly distributed tick values
-                    axis_values = np.linspace(vmin, vmax, 11)
-                    tick_positions = np.linspace(0, colorbar_width, 11)
-
-                    # Prepare data for axis and labels
-                    axis_data = pd.DataFrame(
-                        {"value_axis": axis_values, "x_axis": tick_positions}
-                    )
-
-                    # Create colorbar with linear gradient
-                    colorbar_chart_obj = (
-                        alt.Chart(df_gradient)
-                        .mark_rect(
-                            x=0,
-                            y=0,
-                            width=colorbar_width,
-                            height=colorbar_height,
-                            color=alt.Gradient(
-                                gradient="linear",
-                                stops=[
-                                    alt.GradientStop(color=min_color_str, offset=0),
-                                    alt.GradientStop(color=max_color_str, offset=1),
-                                ],
-                                x1=0,
-                                x2=1,  # Horizontal gradient
-                                y1=0,
-                                y2=0,  # Keep y constant
-                            ),
-                        )
-                        .encode(
-                            x=alt.value(chart_width / 2), y=alt.value(0)
-                        )  # Center colorbar
-                        .properties(width=colorbar_width, height=colorbar_height)
-                    )
-                    # Add tick marks to colorbar
-                    axis_chart = (
-                        alt.Chart(axis_data)
-                        .mark_tick(thickness=2, size=8)
-                        .encode(
-                            x=alt.X("x_axis:Q", axis=None),
-                            y=alt.value(colorbar_height - 2),
-                        )
-                    )
-                    # Add value labels below tick marks
-                    text_labels = (
-                        alt.Chart(axis_data)
-                        .mark_text(baseline="top", fontSize=10, dy=0)
-                        .encode(
-                            x=alt.X("x_axis:Q"),
-                            text=alt.Text("value_axis:Q", format=".1f"),
-                            y=alt.value(colorbar_height + 10),
-                        )
-                    )
-                    # Add title to colorbar
-                    title_chart = (
-                        alt.Chart(pd.DataFrame([{"text_title": layer_name}]))
-                        .mark_text(
-                            fontSize=12,
-                            fontWeight="bold",
-                            baseline="bottom",
-                            align="center",
-                        )
-                        .encode(
-                            text="text_title:N",
-                            x=alt.value(colorbar_width / 2),
-                            y=alt.value(colorbar_height + 40),
-                        )
-                    )
-                    # Combine all colorbar components
-                    combined_colorbar = alt.layer(
-                        colorbar_chart_obj, axis_chart, text_labels, title_chart
-                    ).properties(width=colorbar_width, height=colorbar_height + 50)
-
-                    bar_chart_viz = (
-                        alt.vconcat(bar_chart_viz, combined_colorbar)
-                        .resolve_scale(color="independent")
-                        .configure_view(stroke=None)
-                        if bar_chart_viz is not None
-                        else combined_colorbar
-                    )
-
-            elif colormap:
-                cmap = colormap
-                cmap_scale = alt.Scale(scheme=cmap, domain=[vmin, vmax])
-
-                current_chart = (
-                    alt.Chart(df)
-                    .mark_rect(opacity=alpha)
-                    .encode(
-                        x=alt.X("x:O", axis=None),
-                        y=alt.Y("y:O", axis=None),
-                        color=alt.Color(
-                            "value:Q",
-                            scale=cmap_scale,
-                            title=layer_name,
-                            legend=alt.Legend(title=layer_name)
-                            if portrayal.colorbar
-                            else None,
-                        ),
-                    )
-                    .properties(width=chart_width, height=chart_height)
-                )
-                base = (
-                    alt.layer(current_chart, base)
-                    if base is not None
-                    else current_chart
-                )
-            else:
-                raise ValueError(
-                    f"PropertyLayer {layer_name} portrayal must include 'color' or 'colormap'."
-                )
-        return base, bar_chart_viz
-
-    def draw_propertylayer(
-        self, propertylayer_portrayal: Callable | dict, ax: Axes | None = None
-    ):
-        """Draw property layers on the space."""
+    def draw_propertylayer(self, propertylayer_portrayal, ax: Axes | None = None):
         # Import here to avoid circular imports
         from mesa.visualization.components import PropertyLayerStyle
+
+        ax = ax if ax is not None else self.canvas
 
         def _dict_to_callable(portrayal_dict):
             """Convert legacy dict portrayal to callable."""
@@ -1044,116 +229,15 @@ class SpaceRenderer:
             )
             return ax, None
 
-        if self.backend == "matplotlib":
-            ax = ax if ax is not None else self.ax
-            self.propertylayer_mesh = self._draw_propertylayer_matplotlib(
-                self.space, property_layers, self.propertylayer_portrayal, ax
-            )
-            return self.propertylayer_mesh
-        else:
-            self.propertylayer_mesh = self._draw_propertylayer_altair(
-                self.space, property_layers, self.propertylayer_portrayal
-            )
-            return self.propertylayer_mesh
+        self.propertylayer_mesh = self.backend_renderer.draw_propertylayer(
+            self.space, property_layers, self.propertylayer_portrayal, ax=ax
+        )
+        return self.propertylayer_mesh
 
-    # Main rendering method
-    def render(
-        self,
-        agent_portrayal=None,
-        propertylayer_portrayal=None,
-        ax: Axes | None = None,
-        **kwargs,
-    ):
-        """Render the complete space with structure, agents, and property layers."""
-        structure_kwargs = kwargs.pop("structure_kwargs", {})
-        agent_kwargs = kwargs.pop("agent_kwargs", {})
-
-        if self.backend == "matplotlib":
-            ax = ax if ax is not None else self.ax
-
-            if self.space_mesh is None:
-                self.draw_structure(ax, **structure_kwargs)
-            if agent_portrayal is not None and self.agent_mesh is None:
-                self.draw_agents(agent_portrayal, ax, **agent_kwargs)
-            if propertylayer_portrayal is not None and self.propertylayer_mesh is None:
-                self.draw_propertylayer(propertylayer_portrayal, ax)
-
-            return ax
-        else:
-            # Altair rendering
-            if self.space_mesh is None:
-                self.draw_structure(**structure_kwargs)
-            if agent_portrayal is not None and self.agent_mesh is None:
-                self.draw_agents(agent_portrayal, **agent_kwargs)
-            if propertylayer_portrayal is not None and self.propertylayer_mesh is None:
-                _, pbar = self.draw_propertylayer(propertylayer_portrayal)
-            else:
-                pbar = None
-
-            # Combine all charts
-            charts = [self.space_mesh]
-            if self.agent_mesh:
-                charts.append(self.agent_mesh)
-            if self.propertylayer_mesh:
-                # propertylayer_mesh is a tuple (base, colorbar)
-                charts.append(self.propertylayer_mesh[0])
-
-            # layer all charts
-            final_chart = alt.layer(*charts)
-            # if color bar is present, add it
-            if pbar is not None:
-                final_chart = alt.vconcat(final_chart, pbar).configure_view(stroke=None)
-
-            return final_chart
-
-    @property
-    def canvas(self):
-        """Get the matplotlib axis used for rendering."""
-        if self.backend == "matplotlib":
-            return self.ax
-        else:
-            structure = self.space_mesh if self.space_mesh else None
-            agents = self.agent_mesh if self.agent_mesh else None
-            prop_base, prop_cbar = self.propertylayer_mesh or (None, None)
-
-            spatial_charts_list = [
-                chart for chart in [structure, prop_base, agents] if chart
-            ]
-
-            main_spatial = None
-            if spatial_charts_list:
-                main_spatial = (
-                    spatial_charts_list[0]
-                    if len(spatial_charts_list) == 1
-                    else alt.layer(*spatial_charts_list)
-                )
-
-            # Determine final chart by combining with color bar if present
-            final_chart = None
-            if main_spatial and prop_cbar:
-                final_chart = alt.vconcat(main_spatial, prop_cbar).configure_view(
-                    stroke=None
-                )
-            elif main_spatial:  # Only main_spatial, no prop_cbar
-                final_chart = main_spatial
-            elif prop_cbar:  # Only prop_cbar, no main_spatial
-                final_chart = prop_cbar
-                final_chart = final_chart.configure_view(grid=False)
-
-            if final_chart is None:
-                # If no charts are available, return an empty chart
-                final_chart = (
-                    alt.Chart(pd.DataFrame())
-                    .mark_point()
-                    .properties(width=450, height=350)
-                )
-
-            final_chart = final_chart.configure_view(stroke="black", strokeWidth=1.5)
-
-            return final_chart
+    def render(self, agent_portrayal=None, propertylayer_portrayal=None, **kwargs):
+        return self.backend_renderer.render(
+            agent_portrayal, propertylayer_portrayal, **kwargs
+        )
 
     def clear_meshes(self):
-        """Clear all stored mesh objects."""
-        self.space_mesh = None
-        self.agent_mesh = None
-        self.propertylayer_mesh = None
+        self.backend_renderer.clear_meshes()
